Add failing tests to simulate the after-query callback failure
When the publisher from `Statement#execute` completes while publisher
from `Result#map` or `Result#getRowsUpdated`, the after query callback
is called.
The after callback should be invoked when the publisher from the
`Result` operations have finished.
This commit adds tests to simulate this situation and verifies it fails
with current code.

[#94]

Signed-off-by: Tadaya Tsuyukubo <tadaya@ttddyy.net>