Trigger `afterQuery` on default methods in `Result`
Prior to this change, the `afterQuery` method was only invoked with
`Result#map(BiFunction)` and not on default methods,
`Result#map(Function)` and `Result#flatMap`. This limitation arose from
the default methods being defined on the `Result` interface.
When a `Result` proxy was called with these default methods, the proxy
logic simply delegated the call to the original instance, which
performs the invocation of the default method. As a result, no
transformation occurred on the mapping method argument and return
values from default methods.

This commit addresses the issue by adding proper handling for the
default methods. It includes returned publisher transformation for
invoking `afterQuery`, proxy creation for `Row/RowSegment`, and
introduces a new `RowSegment` proxy.

Fixes #138