Use static interface list in JdkProxyFactory
In `1.1.3`, `JdkProxyFactory` changed to use `HashSet` to specify the
list of interfaces for making proxies on row proxies.  While this works
for the regular jvm, it has an issue for graalvm, which requires a
static list of interfaces.
In this commit, the list of interfaces has been reverted to a static
list As a result, the generated proxy object for `Row` cannot be cast
to the implementation specific interfaces. However, it remains possible
to unwrap to the implementation class using `Wrapped#unwrap`.

Fixes #142